Move bounce logic to concern
