+++ config/application.rb
@@ -10,6 +10,7 @@ module RestWhois
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 6.0
+    config.autoloader = :classic
 
     # Authorize all hosts
     config.hosts.clear
